# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Khadamatak SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List opportunities
      parameters:
        - name: stage
          in: query
          schema:
            type: string
            enum: [new, qualified, proposal, negotiation, won, lost]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Opportunities list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunitiesListResponse"
    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Create opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
  /opportunities/{id}:
    patch:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpportunityBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityResponse"
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, rejected, cancelled]
        - name: method
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentsListResponse"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create payment (manual, starts as pending)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
...
components:
  schemas:
...
    Opportunity:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        stage:
          type: string
          enum: [new, qualified, proposal, negotiation, won, lost]
        value:
          type: number
        currency:
          type: string
          default: "YER"
        contactName:
          type: string
        probability:
          type: integer
        expectedCloseDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, title, stage]
    OpportunityResponse:
      type: object
      properties:
        opportunity:
          $ref: "#/components/schemas/Opportunity"
    OpportunitiesListResponse:
      type: object
      properties:
        opportunities:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
        total:
          type: integer
        pipelineValue:
          type: number
        stageCounts:
          type: object
          additionalProperties:
            type: integer
    CreateOpportunityBody:
      type: object
      properties:
        title:
          type: string
        value:
          type: number
        stage:
          type: string
          enum: [new, qualified, proposal, negotiation, won, lost]
        contactId:
          type: string
        probability:
          type: integer
        expectedCloseDate:
          type: string
          format: date
      required: [title, stage]
    UpdateOpportunityBody:
      type: object
      properties:
        title:
          type: string
        stage:
          type: string
          enum: [new, qualified, proposal, negotiation, won, lost]
        value:
          type: number
        probability:
          type: integer
...
    Payment:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: "YER"
        method:
          type: string
          enum: [cash, kuraimi, jawali, bank_transfer, other]
        status:
          type: string
          enum: [pending, confirmed, rejected, cancelled]
        contactName:
          type: string
        reference:
          type: string
        notes:
          type: string
        confirmedByName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, amount, method, status]
    PaymentResponse:
      type: object
      properties:
        payment:
          $ref: "#/components/schemas/Payment"
    PaymentsListResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        totalConfirmed:
          type: number
        totalPending:
          type: number
    CreatePaymentBody:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [cash, kuraimi, jawali, bank_transfer, other]
        contactId:
          type: string
        orderId:
          type: string
        reference:
          type: string
        notes:
          type: string
      required: [amount, method]
...